pyproject.toml
[project]
name = "odoo-tools"
authors = [
    {name = "Camptocamp SA", email="info@camptocamp.com"},
]
description = "Our Odoo project tools"
readme = "README.md"
keywords = ["odoo", "project"]
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Framework :: Odoo",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]
dependencies = [
    "invoke==2.0.0",
    "requests",
    "git-aggregator",
    "git-autoshare",
    "passlib",
    "clipboard",
    "marabunta",
    "pre-commit",
    "psycopg2-binary>=2.7.6",
    "gitpython",
    "pyyaml>6.0.0",
    "ruamel.yaml",
    "kaptan",
    "wheel",
    "cookiecutter",
    "click",
    "towncrier",
    "importlib_resources; python_version < '3.9'",
    "bump2version",
    # Include fix for parsing editable requirements
    # https://github.com/madpah/requirements-parser/pull/85
    "requirements-parser @ https://github.com/simahawk/requirements-parser/archive/refs/heads/fix-editable.tar.gz"
]
requires-python = ">=3.8"
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/camptocamp/odoo-project-tools"
documentation = "https://github.com/camptocamp/odoo-project-tools"
repository = "https://github.com/camptocamp/odoo-project-tools"

[project.scripts]
otools-tasks = "odoo_tools.tasks.main:program.run"
otools-project = "odoo_tools.project:cli"
otools-addon = "odoo_tools.addon:cli"
otools-release = "odoo_tools.release:cli"

[project.optional-dependencies]
test = [
  "coverage",
  "pytest",
  "responses",
]

[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}

[tool.setuptools_scm]


[tool.black]
line-length = 88
skip-string-normalization = "True"

[tool.towncrier]
directory = "changes.d"
filename = "HISTORY.rst"
template = ".towncrier-template.rst"
title_format = ""
issue_format = "{issue}"

[[tool.towncrier.type]]
directory = "feature"
name = "Features and Improvements"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bugfixes"
showcontent = true

[[tool.towncrier.type]]
directory = "doc"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "build"
name = "Build"
showcontent = true
